refactor(rom): new image strategy menuimage
